src/python-ml/requirements.txt
numpy
scikit-learn
hdbscan
pandas
umap-learn
